+++ packages/bluesky/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.15.6
 
 - Removed `getTimelineSkeleton` from `UnspeccedService`. ([#1226](https://github.com/myConsciousness/atproto.dart/issues/1226))
+- Added `getTaggedSuggestions` on `UnspeccedService`. ([#1225](https://github.com/myConsciousness/atproto.dart/issues/1225))
 
 ## v0.15.5